fix email template

+++ public/partials/mwb-wpr-points-template.php
@@ -98,7 +98,8 @@ if ( ! is_array( $coupon_settings ) ) {
 	<div class="mwb_wpr_points_only"><p class="mwb_wpr_heading_para" >
 		<span class="mwb_wpr_heading"><?php echo esc_html( $mwb_text_points_value ) . ':'; ?></span></p>
 		<?php
-		$get_points = (int) get_user_meta( $user_id, 'mwb_wpr_points', true );
+		// $get_points = (int) get_user_meta( $user_id, 'mwb_wpr_points', true );
+		$get_points = get_user_meta( $user_id, 'mwb_wpr_points', true );
 		$get_point = get_user_meta( $user_id, 'points_details', true );
 		?>
 		<span class="mwb_wpr_heading" id="mwb_wpr_points_only">
